# api/docs/swagger.yaml
definitions:
  healthAnalytics.AddMedicalRecordRequest:
    properties:
      attachments:
        items:
          type: string
        type: array
      description:
        type: string
      doctorId:
        type: string
      recordDate:
        type: string
      recordType:
        type: string
      userId:
        type: string
    type: object
  healthAnalytics.AllLifestyles:
    properties:
      lifestyles:
        items:
          $ref: '#/definitions/healthAnalytics.LifestyleResponse'
        type: array
    type: object
...
  healthAnalytics.Lifestyle:
    properties:
      dataType:
        type: string
      dataValue:
        type: string
      recordedData:
        type: string
      userId:
        type: string
    type: object
  healthAnalytics.LifestyleResponse:
    properties:
      createdAt:
        type: string
      dataType:
        type: string
      dataValue:
        type: string
      id:
        type: string
      recordedData:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
...
  healthAnalytics.ListMedicalRecord:
    properties:
      medicalRecords:
        items:
          $ref: '#/definitions/healthAnalytics.MedicalRecord'
        type: array
    type: object
  healthAnalytics.MedicalRecord:
    properties:
      attachments:
        items:
          type: string
        type: array
      createdAt:
        type: string
      description:
        type: string
      doctorId:
        type: string
      id:
        type: string
      recordDate:
        type: string
      recordType:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  healthAnalytics.MedicalRecordID:
    properties:
      id:
        type: string
    type: object
...
  healthAnalytics.Message:
    properties:
      message:
        type: string
    type: object
...
  healthAnalytics.UpdateLifestyle:
    properties:
      dataType:
        type: string
      dataValue:
        type: string
      id:
        type: string
    type: object
  healthAnalytics.UpdateMedicalRecordReq:
    properties:
      id:
        type: string
      recordDate:
        type: string
      recordType:
        type: string
    type: object
...
  models.Error:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: API-Gateway service
  version: "1.0"
paths:
...
  /api/lifestyle:
    get:
      consumes:
      - application/json
      description: Lists all lifestyle data with optional filters
      parameters:
      - description: Pagination limit
        in: query
        name: limit
        type: string
      - description: Pagination offset
        in: query
        name: offset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.AllLifestyles'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: List Lifestyle Data
      tags:
      - Lifestyle
    post:
      consumes:
      - application/json
      description: Adds new lifestyle data for the user
      parameters:
      - description: Lifestyle Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/healthAnalytics.Lifestyle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.LifestyleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Add Lifestyle Data
      tags:
      - Lifestyle
  /api/lifestyle/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes lifestyle data by ID
      parameters:
      - description: Lifestyle Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete Lifestyle Data
      tags:
      - Lifestyle
    get:
      consumes:
      - application/json
      description: Retrieves lifestyle data by ID
      parameters:
      - description: Lifestyle Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.LifestyleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Lifestyle Data
      tags:
      - Lifestyle
    put:
      consumes:
      - application/json
      description: Updates existing lifestyle data by ID
      parameters:
      - description: Lifestyle Data ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Lifestyle Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/healthAnalytics.UpdateLifestyle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.LifestyleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update Lifestyle Data
      tags:
      - Lifestyle
  /api/medical-records:
    get:
      consumes:
      - application/json
      description: Lists medical records based on filters
      parameters:
      - description: Pagination offset
        in: query
        name: offset
        type: string
      - description: Pagination limit
        in: query
        name: limit
        type: string
      - description: Description filter
        in: query
        name: description
        type: string
      - description: Doctor ID filter
        in: query
        name: doctor_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.ListMedicalRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: List Medical Records
      tags:
      - MedicalRecords
    post:
      consumes:
      - application/json
      description: Adds a new medical record
      parameters:
      - description: Medical Record
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/healthAnalytics.AddMedicalRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.MedicalRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Add Medical Record
      tags:
      - MedicalRecords
  /api/medical-records/{id}:
    delete:
      description: Deletes a medical record by ID
      parameters:
      - description: Medical Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.MedicalRecordID'
        "400":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete Medical Record
      tags:
      - MedicalRecords
    get:
      description: Retrieves a medical record by ID
      parameters:
      - description: Medical Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.MedicalRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Medical Record
      tags:
      - MedicalRecords
    put:
      consumes:
      - application/json
      description: Updates an existing medical record
      parameters:
      - description: Medical Record ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Medical Record
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/healthAnalytics.UpdateMedicalRecordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.MedicalRecord'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update Medical Record
      tags:
      - MedicalRecords
...
schemes:
- http
swagger: "2.0"